Brought in Thanuj's work on functional component
